+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/ImplementationBridgeHelpers.java
@@ -1053,9 +1053,11 @@ public class ImplementationBridgeHelpers {
                 int targetedCountAfterSplit);
 
             String getLinkWithoutTrailingSlash(CosmosAsyncContainer cosmosAsyncContainer);
+            Mono<String> extractCollectionRid(CosmosAsyncContainer cosmosAsyncContainer);
             Mono<Boolean> checkFeedRangeOverlapping(CosmosAsyncContainer container, FeedRange feedRange1, FeedRange feedRange2);
             Mono<List<FeedRange>> getOverlappingFeedRanges(CosmosAsyncContainer container, FeedRange feedRange, boolean forceRefresh);
             Mono<PartitionKeyDefinition> getPartitionKeyDefinition(CosmosAsyncContainer container);
+            String getLinkWithoutTrailingSlash(String link);
         }
     }